extensions REFACTOR cleanup public header
The plugins_exts.h is supposed to be public, so move the unnecessary
structures into the internal extensions header.
