+++ src/plugins_exts.c
@@ -13,6 +13,7 @@
  */
 
 #include "plugins_exts.h"
+#include "plugins_exts_internal.h"
 
 #include <stdint.h>
 #include <string.h>
@@ -24,7 +25,7 @@
 /**
  * @brief list of all extension plugins implemented internally
  */
-struct lyext_plugins_list lyext_plugins_internal[] = {
+struct lyext_plugin_record lyext_plugins_internal[] = {
     {"ietf-netconf-acm", "2012-02-22", "default-deny-write", &nacm_plugin},
     {"ietf-netconf-acm", "2018-02-14", "default-deny-write", &nacm_plugin},
     {"ietf-netconf-acm", "2012-02-22", "default-deny-all", &nacm_plugin},
