+++ src/plugins_exts_internal.h
@@ -21,9 +21,23 @@
 #include "plugins_exts_metadata.h"
 
 /**
+ * @brief Record describing an implemented extension.
+ */
+struct lyext_plugin_record {
+    const char *module;          /**< name of the module where the extension is defined */
+    const char *revision;        /**< optional module revision - if not specified, the plugin applies to any revision,
+                                      which is not an optimal approach due to a possible future revisions of the module.
+                                      Instead, there should be defined multiple items in the plugins list, each with the
+                                      different revision, but all with the same pointer to the plugin extension. The
+                                      only valid use case for the NULL revision is the case the module has no revision. */
+    const char *name;            /**< name of the extension */
+    struct lyext_plugin *plugin; /**< plugin for the extension */
+};
+
+/**
  * @brief List of internally implemented extension plugins.
  */
-extern struct lyext_plugins_list lyext_plugins_internal[];
+extern struct lyext_plugin_record lyext_plugins_internal[];
 
 /**
  * @brief Index of Metadata's annotation extension plugin in lyext_plugins_internal
